Move orion_metrics_rollup to common_policies
The rollup job only reads from and writes to Elasticsearch and has no
per-resource or per-region AWS behaviour, so it belongs with the other
cloud-agnostic common policies rather than under policy/aws. Dispatch it
through the COMMON_POLICIES path instead of cost_policies, and drop the
now-dead auto-discovery exclusion (get_policies only scans policy/aws).
